% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth_charts_data.r
\docType{data}
\name{curves}
\alias{curves}
\title{Data from growth charts}
\format{
Data frame with 14 variables and 436 observations:
\describe{
\item{Gender}{A character string. Refers to Male or Female.}
\item{Agemos}{Age in months}
\item{Power}{Power in the Box-Cox transformation}
\item{Median}{Median}
\item{CV}{Generalized Coeficient of Variation}
\item{P3}{Data under percentile 3}
\item{P5}{Data under percentile 5}
\item{P10}{Data under percentile 10}
\item{P25}{Data under percentile 25}
\item{P50}{Data under percentile 50}
\item{P75}{Data under percentile 75}
\item{P90}{Data under percentile 90}
\item{P95}{Data under percentile 95}
\item{P97}{Data under percentile 97 }

}
}
\usage{
curves
}
\description{
Data from growth charts
}
\details{
A data frame containing information for growth charts from the National Center from Health and Statistics. Visit \url{https://www.cdc.gov/growthcharts/percentile_data_files.htm} for more details.
}
\keyword{datasets}
