% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_sample.R
\docType{data}
\name{data_sample}
\alias{data_sample}
\title{A sample dataset for demostration purposes.}
\format{
Data frame with 13 variables and 20 observations:
\describe{
\item{Athlete}{A character string. Name of the subjects.}
\item{\verb{Date of Birth}}{A date object referring to the DOB for each athlete.}
\item{\verb{Testing Date}}{A date object referring to the date of testing of each athlete.}
\item{Gender}{A character string. Refers to Male or Female.}
\item{\code{Weight1 (KG)}}{A number. Weight in kgs. Measurement 1.}
\item{\code{Weight2 (KG)}}{A number. Weight in kgs. Measurement 2.}
\item{\code{Height1 (CM)}}{A number. Height in cms. Measurement 1.}
\item{\code{Height2 (CM)}}{A number. Height in cms. Measurement 2.}
\item{\verb{Sitting Height1 (CM)}}{A number. Length of the trunk in cms for a seated measurement. Measurement 1.}
\item{\verb{Sitting Height2 (CM)}}{A number. Length of the trunk in cms for a seated measurement. Measurement 2.}
\item{\verb{Bench Height2 (CM)}}{A number. If the sitting height is done using a chair or a bench, indicate its length, otherwise use 0.}
\item{\verb{Mothers Height (CM)}}{A number. The standing height of the athlete's mother in cms.}
\item{\verb{Fathers Height (CM)}}{A number. The standing height of the athlete's father in cms.}

}
}
\usage{
data_sample
}
\description{
A sample dataset for demostration purposes.
}
\details{
A data frame containing unreal sample data for demostration purposes only. It also serves as an example for a template of how user collected data should look.
Use this dataset to learn about the functions on this package.
}
\keyword{datasets}
