% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_maturity_offset.R
\name{plot_maturity_offset}
\alias{plot_maturity_offset}
\title{Maturity Offset Plot}
\usage{
plot_maturity_offset(data)
}
\arguments{
\item{data}{A data frame. The object containing the raw data we wish to analize.}
}
\value{
A lolliplot plot \code{(ggplot)}
}
\description{
This function returns a lollipop ggplot object showing the offset in years from current age to estimated age at PHV for each athlete in the dataset.
}
\details{
Refer to references cited on this package for further details on how these metrics are calculated.
}
\examples{
plot_maturity_offset(data_sample)

}
