% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_predicted_height_cm.R
\name{plot_predicted_height_cm}
\alias{plot_predicted_height_cm}
\title{Predicted Height Plot (cms)}
\usage{
plot_predicted_height_cm(data)
}
\arguments{
\item{data}{A data frame. The object containing the raw data we wish to analize.}
}
\value{
A plot \code{(ggplot)}
}
\description{
This function returns a ggplot object showing the predicted adult height for each athlete in the dataset.
For the same plot in inches use \strong{\code{plot_predicted_height_in()}}
}
\examples{
plot_predicted_height_cm(data_sample)

}
