% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_predicted_height_in.R
\name{plot_predicted_height_in}
\alias{plot_predicted_height_in}
\title{Predicted Height Plot (Inches)}
\usage{
plot_predicted_height_in(data)
}
\arguments{
\item{data}{A data frame. The object containing the raw data we wish to analize.}
}
\value{
A plot \code{(ggplot)}
}
\description{
This function returns a ggplot object showing the predicted adult height in inches for each athlete in the dataset.
For the same plot in centimeters use \strong{\code{plot_predicted_height_cm()}}
}
\examples{
plot_predicted_height_in(data_sample)

}
