% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_puberty_stages.R
\name{plot_puberty_stages}
\alias{plot_puberty_stages}
\title{\% Adult Height vs Maturity Offset in Years}
\usage{
plot_puberty_stages(data)
}
\arguments{
\item{data}{A data frame. The object containing the raw data we wish to analize.}
}
\value{
A plot \code{(ggplot)}
}
\description{
This function returns a scatterplot showing the \% of adult height vs the maturity offset \verb{(in years)}.
}
\examples{
plot_puberty_stages(data_sample)

}
