% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_time_phv.R
\name{plot_time_phv}
\alias{plot_time_phv}
\title{Time to PHV Dumbell Plot}
\usage{
plot_time_phv(data)
}
\arguments{
\item{data}{A data frame. The object containing the raw data we wish to analize.}
}
\value{
A dumbell plot \code{(ggplot)}
}
\description{
This function returns a dumbell plot showing the difference (in years) between current age and estimated age at PHV for each athlete in the dataset.
}
\details{
Athletes are ordered by the difference between current and estimated age at PHV, as shown on the right of the plot, from highest to lowest.

Check the references cited on this package for further details on how these metrics are calculated.
}
\examples{
plot_time_phv(data_sample)

}
