% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_puberty_stages.R
\name{plot_puberty_stages}
\alias{plot_puberty_stages}
\title{\% Adult Height & Puberty States Plot}
\usage{
plot_puberty_stages(data)
}
\arguments{
\item{data}{A data frame. The object containing the raw data we wish to analize.}
}
\value{
A plot \code{(ggplot)}
}
\description{
This function returns a ggplot object showing the \% of adult height for each subject in the dataset. It adds a background color to highlight maturity categories.
}
\examples{
plot_puberty_stages(data_sample)

}
